Ulysses Lee "Junior" Bridgeman
Junior Bridgeman was a 6'5" All American guard for the University of
Louisville who graduated with a degree in psychology in 1979 and was
later inducted into the UofL Athletics Hall of Fame.

After playing for UofL, Bridgeman played 12 years in the NBA for the
Milwaukee Bucks and the LA Clippers serving as the president of the NBA
Player's Association.  The Bucks retired his jersey in 1988 and he was
inducted into the Wisconsin Athletics Hall of Fame.  Following his
basketball career Bridgeman became a successful businessman whose net
worth is estimated to be in the hundreds of millions.  He currently
serves as CEO of his own company which employs over 10000 people and
serves on several boards of directors of other companies.

I got his autograph in 2004 when, as a member of the University of
Louisville's Board of Trustees, he signed my Bachelor of Science degree
:).

http://louisville.edu/artsandsciences/about/hallofhonor/inductees/bridgeman.html